pyproject.toml
[tool.poetry]
name = "refvision"
version = "0.1.0"
description = "RefVision AI for powerlifting"
authors = ["Michael Kingston <71872141+mkingopng@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "^2.5.1"
torchvision = "^0.20.1"
ultralytics = "^8.3.27"
ruff = "^0.7.4"
pytest = "^8.3.3"
black = "^24.10.0"
moviepy = "^1.0.3"
python-dotenv = "^1.0.1"
inference-cli = "^0.27.0"
boto3 = "^1.35.73"
aws-cdk-lib = "^2.173.2"
constructs = "^10.4.2"
lap = "^0.5.12"
flask = "^3.1.0"
gunicorn = "^23.0.0"
pytest-cov = "^6.0.0"
mypy = "^1.15.0"
rsa = "^4.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest-mock = "^3.14.0"
types-pyyaml = "^6.0.12.20241230"
boto3-stubs = {extras = ["essential"], version = "^1.36.19"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "cdk/cdk\\.out"

[tool.mypy.yaml]
ignore_missing_imports = true
follow_imports = "skip"

[tool.ruff.per-file-ignores]
"ref_vision_infrastructure_stack.py" = ["F401", "F841"]

[tool.poetry.scripts]
run-pipeline = "refvision.scripts.run_pipeline:main"
